crud of unit and UI of lesson listing in edit page of unit

+++ app/Http/Controllers/CourseController.php
@@ -6,16 +6,17 @@ use App\Models\Category;
 use App\Models\Course;
 use App\Models\Image;
 use App\Models\Level;
-use App\Models\Unit;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
 class CourseController extends Controller
 {
-    public function index(Category $category, Level $level) {
-
+    public function index(Category $category, Level $level)
+    {
         return view('courses.index', [
-            'courses' => Course::filter(request(['search', 'category', 'orderBy', 'level']))->paginate(10),
+            'courses' => Course::courseOwner()
+                ->filter(request(['search', 'category', 'orderBy', 'level']))
+                ->paginate(),
             'categories' => $category->get(),
             'currentCategory' => $category->find(request()->category),
             'levels' => $level->get(),
@@ -23,23 +24,16 @@ class CourseController extends Controller
         ]);
     }
 
-    public function show(Course $course) {
-        return view('courses.units.index', [
-            'course' => $course,
-            'units' => Unit::get(),
-        ]);
-    }
-    
-    public function create() {
-
+    public function create()
+    {
         return view('courses.create', [
             'categories' => Category::get(),
             'levels' => Level::get(),
         ]);
     }
 
-    public function store(Request $request) {
-
+    public function store(Request $request)
+    {
         $attributes = $request->validate([
             'title' => ['required','min:3', 'max:50'],
             'description' => ['required', 'min:5', 'max:255'],
@@ -53,16 +47,20 @@ class CourseController extends Controller
             ],
         ]);
 
-        $attributes += ['user_id' => Auth::id()];
-        
+        $attributes += [
+            'user_id' => Auth::id()
+        ];
+
         Course::create($attributes);
 
         return back()->with('success', 'Course created successfully.');
     }
 
-    public function edit(Course $course) {
+    public function edit(Course $course)
+    {
+        $this->authorize('edit', $course);
 
-        return view('courses.edit', [
+        return view('courses._course_information', [
             'course' => $course,
             'categories' => Category::get(),
             'levels' => Level::get(),
@@ -70,8 +68,20 @@ class CourseController extends Controller
         ]);
     }
 
-    public function update(Request $request, Course $course) {
-        
+    public function show(Course $course)
+    {
+        $this->authorize('show', $course);
+
+        return view('courses.units.index', [
+            'course' => $course,
+            'units' => $course->units()->get(),
+        ]);
+    }
+
+    public function update(Request $request, Course $course)
+    {
+        $this->authorize('update', $course);
+
         $attributes = $request->validate([
             'title' => ['required', 'min:3', 'max:50'],
             'description' => ['required', 'min:5', 'max:255'],
@@ -86,21 +96,23 @@ class CourseController extends Controller
         ]);
 
         ($request->certificate == 1)? $attributes += ['certificate' => 1] : $attributes += ['certificate' => 0];
-        
+
         $course->update($attributes);
-        
+
         if(request()->file('image')) {
-            $path = $request->image->store('public/images');
+            $path = $request->image->store('images');
             Image::create([
                 'image_path' => $path,
                 'course_id' => $course->id
             ]);
         }
 
-        return redirect()->route('courses.index')->with('success', 'Course updated successfully.');
+        return back()->with('success', 'Course updated successfully.');
     }
 
-    public function delete(Course $course) {
+    public function delete(Course $course)
+    {
+        $this->authorize('delete', $course);
         $course->delete();
 
         return back()->with('success', 'Course deleted successfully.');
